+++ activerecord/test/models/bulb.rb
@@ -37,3 +37,9 @@ class CustomBulb < Bulb
     self.frickinawesome = true if name == 'Dude'
   end
 end
+
+class FunkyBulb < Bulb
+  before_destroy do
+    raise "before_destroy was called"
+  end
+end